refactor(enrollment): collapse EnrollmentPolicy to schema-as-struct
Slice 1 of the Enrollment flatten. Merge domain model + Ecto schema +
mapper + repository + 2 ports into KlassHero.Enrollment.EnrollmentPolicy,
and inline the policy use cases (upsert + capacity reads) as functions on
the context calling Repo directly. set_enrollment_policy is wrapped in
context_span; reads stay bare.

- New: lib/klass_hero/enrollment/enrollment_policy.ex (schema + changeset
  gatekeeper + pure capacity math)
- EnrollmentRepository capacity check repointed at the new module (mapper
  indirection dropped)
- seeds + config policy DI + interaction_sweep probe updated
- Test tree flattened: enrollment_policy_test.exs replaces the deleted
  model/schema/mapper/repo/query unit tests
